update MVVM for grid
